add handler in library
